Improve single file performance - raw and csv files (#25)
Optimize CsvFileIngestService and RawFileIngestService.

CSV Events are now batched in a list and entire batch is converted to json. Each record is parsed into a hashmap before being added to the batch.
For Raw file events, BufferedInputStream is used now
Also the service class names in test scripts are fixed and new script is added for testing csv files.